admin, user api 확인, 대여, 반납 api 구현

+++ back/src/main/java/playbook/encore/back/controller/SortFirstController.java
@@ -1,5 +1,6 @@
 package playbook.encore.back.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.graphql.GraphQlProperties;
 import org.springframework.http.HttpStatus;
@@ -8,6 +9,7 @@ import org.springframework.web.bind.annotation.*;
 import playbook.encore.back.data.dao.SortFirstDAO;
 import playbook.encore.back.data.dto.sortFirst.SortFirstRequestDto;
 import playbook.encore.back.data.dto.sortFirst.SortFirstResponseDto;
+import playbook.encore.back.interceptor.LoginCheckInterceptor;
 import playbook.encore.back.service.SortFirstService;
 
 import java.util.List;
@@ -17,12 +19,10 @@ import java.util.List;
 public class SortFirstController {
 
     private final SortFirstService sortFirstService;
-    private final SortFirstDAO sortFirstDAO;
 
     @Autowired
-    public SortFirstController(SortFirstService sortFirstService, SortFirstDAO sortFirstDAO) {
+    public SortFirstController(SortFirstService sortFirstService) {
         this.sortFirstService = sortFirstService;
-        this.sortFirstDAO = sortFirstDAO;
     }
 
     @GetMapping
@@ -32,23 +32,39 @@ public class SortFirstController {
     }
 
     @PostMapping
-    public ResponseEntity<SortFirstResponseDto> postSortFirst(@RequestBody SortFirstRequestDto sortFirstRequestDto) throws Exception {
-        SortFirstResponseDto sortFirstResponseDto = sortFirstService.insertSortFirst(sortFirstRequestDto);
-        return ResponseEntity.status(HttpStatus.CREATED).body(sortFirstResponseDto);
+    public ResponseEntity<?> postSortFirst(
+            HttpServletRequest request,
+            @RequestBody SortFirstRequestDto sortFirstRequestDto
+    ) throws Exception {
+        if (request.getAttribute("ROLE") == LoginCheckInterceptor.RoleType.ADMIN) {
+            SortFirstResponseDto sortFirstResponseDto = sortFirstService.insertSortFirst(sortFirstRequestDto);
+            return ResponseEntity.status(HttpStatus.CREATED).body(sortFirstResponseDto);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("관리자만 접근 가능합니다.");
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<SortFirstResponseDto> putSortFirstById(
+    public ResponseEntity<?> putSortFirstById(
+            HttpServletRequest request,
             @PathVariable("id") Integer sortFirstId,
             @RequestBody SortFirstRequestDto sortFirstRequestDto
     ) throws Exception{
-        SortFirstResponseDto sortFirstResponseDto = sortFirstService.changeSortFirst(sortFirstId, sortFirstRequestDto);
-        return ResponseEntity.status(HttpStatus.OK).body(sortFirstResponseDto);
+        if (request.getAttribute("ROLE") == LoginCheckInterceptor.RoleType.ADMIN) {
+            SortFirstResponseDto sortFirstResponseDto = sortFirstService.changeSortFirst(sortFirstId, sortFirstRequestDto);
+            return ResponseEntity.status(HttpStatus.OK).body(sortFirstResponseDto);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("관리자만 접근 가능합니다.");
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteSortFirst(@PathVariable("id") Integer sortFirstId) throws Exception {
-        sortFirstService.deleteSortFirstById(sortFirstId);
-        return ResponseEntity.status(HttpStatus.OK).body("삭제를 수행하였습니다.");
+    public ResponseEntity<String> deleteSortFirst(
+            HttpServletRequest request,
+            @PathVariable("id") Integer sortFirstId
+    ) throws Exception {
+        if (request.getAttribute("ROLE") == LoginCheckInterceptor.RoleType.ADMIN) {
+            sortFirstService.deleteSortFirstById(sortFirstId);
+            return ResponseEntity.status(HttpStatus.OK).body("삭제를 수행하였습니다.");
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("관리자만 접근 가능합니다.");
     }
 }
