+++ back/src/main/java/playbook/encore/back/controller/SortSecondController.java
@@ -1,5 +1,6 @@
 package playbook.encore.back.controller;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,6 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import playbook.encore.back.data.dao.SortSecondDAO;
 import playbook.encore.back.data.dto.sortSecond.SortSecondRequestDto;
 import playbook.encore.back.data.dto.sortSecond.SortSecondResponseDto;
+import playbook.encore.back.interceptor.LoginCheckInterceptor;
 import playbook.encore.back.service.SortSecondService;
 
 import java.util.List;
@@ -16,12 +18,10 @@ import java.util.List;
 public class SortSecondController {
 
     private final SortSecondService sortSecondService;
-    private final SortSecondDAO sortSecondDAO;
 
     @Autowired
-    public SortSecondController(SortSecondService sortSecondService, SortSecondDAO sortSecondDAO) {
+    public SortSecondController(SortSecondService sortSecondService) {
         this.sortSecondService = sortSecondService;
-        this.sortSecondDAO = sortSecondDAO;
     }
 
     @GetMapping
@@ -31,24 +31,43 @@ public class SortSecondController {
     }
 
     @PostMapping
-    public ResponseEntity<SortSecondResponseDto> postSortSecond(@RequestBody SortSecondRequestDto sortSecondRequestDto) throws Exception {
-        SortSecondResponseDto sortSecondResponseDto = sortSecondService.insertSortSecond(sortSecondRequestDto);
-        return ResponseEntity.status(HttpStatus.CREATED).body(sortSecondResponseDto);
+    public ResponseEntity<?> postSortSecond(
+            HttpServletRequest request,
+            @RequestBody SortSecondRequestDto sortSecondRequestDto
+    ) throws Exception {
+        if (request.getAttribute("ROLE") == LoginCheckInterceptor.RoleType.ADMIN) {
+            SortSecondResponseDto sortSecondResponseDto = sortSecondService.insertSortSecond(sortSecondRequestDto);
+            return ResponseEntity.status(HttpStatus.CREATED).body(sortSecondResponseDto);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("관리자만 접근 가능합니다.");
+
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<SortSecondResponseDto> putSortSecondById(
+    public ResponseEntity<?> putSortSecondById(
+            HttpServletRequest request,
             @PathVariable("id") Integer sortSecondId,
             @RequestBody SortSecondRequestDto sortSecondRequestDto
     ) throws Exception{
-        SortSecondResponseDto sortSecondResponseDto = sortSecondService.changeSortSecond(sortSecondId, sortSecondRequestDto);
-        return ResponseEntity.status(HttpStatus.OK).body(sortSecondResponseDto);
+        if (request.getAttribute("ROLE") == LoginCheckInterceptor.RoleType.ADMIN) {
+            SortSecondResponseDto sortSecondResponseDto = sortSecondService.changeSortSecond(sortSecondId, sortSecondRequestDto);
+            return ResponseEntity.status(HttpStatus.OK).body(sortSecondResponseDto);
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("관리자만 접근 가능합니다.");
+
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteSortSecond(@PathVariable("id") Integer sortSecondId) throws Exception {
-        sortSecondService.deleteSortSecondById(sortSecondId);
-        return ResponseEntity.status(HttpStatus.OK).body("삭제를 수행하였습니다.");
+    public ResponseEntity<String> deleteSortSecond(
+            HttpServletRequest request,
+            @PathVariable("id") Integer sortSecondId
+    ) throws Exception {
+        if (request.getAttribute("ROLE") == LoginCheckInterceptor.RoleType.ADMIN) {
+            sortSecondService.deleteSortSecondById(sortSecondId);
+            return ResponseEntity.status(HttpStatus.OK).body("삭제를 수행하였습니다.");
+        }
+        return ResponseEntity.status(HttpStatus.FORBIDDEN).body("관리자만 접근 가능합니다.");
+
     }
 
 }
